# api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "rusunawa.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/bookings": {
      "get": {
        "operationId": "BookingService_GetBookings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetBookingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BookingService"
        ]
      },
      "post": {
        "operationId": "BookingService_CreateBooking",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseCreateBookingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/baseCreateBookingRequest"
            }
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
    "/v1/bookings/{bookingId}": {
      "get": {
        "operationId": "BookingService_GetBooking",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetBookingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BookingService"
        ]
      }
    },
...
    "/v1/invoices": {
      "get": {
        "operationId": "PaymentService_GetInvoices",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetInvoicesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "status",
            "description": "e.g., \"paid\", \"pending\", \"overdue\"",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "PaymentService"
        ]
      },
      "post": {
        "summary": "Existing invoice methods",
        "operationId": "PaymentService_CreateInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseCreateInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/baseCreateInvoiceRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    },
    "/v1/invoices/{invoiceId}": {
      "get": {
        "operationId": "PaymentService_GetInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "invoiceId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    },
...
    "/v1/issues": {
      "get": {
        "operationId": "IssueService_GetIssues",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetIssuesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "IssueService"
        ]
      },
      "post": {
        "operationId": "IssueService_ReportIssue",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseReportIssueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/baseReportIssueRequest"
            }
          }
        ],
        "tags": [
          "IssueService"
        ]
      }
    },
    "/v1/issues/{issueId}": {
      "get": {
        "operationId": "IssueService_GetIssue",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetIssueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "issueId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "IssueService"
        ]
      }
    },
...
    "/v1/payments/callback": {
      "post": {
        "summary": "Midtrans callback handler",
        "operationId": "PaymentService_HandlePaymentCallback",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseMessageStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/basePaymentCallbackRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    },
...
    "/v1/rooms": {
      "get": {
        "operationId": "RoomService_GetRooms",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetRoomsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "rentalType",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "RoomService"
        ]
      },
      "post": {
        "operationId": "RoomService_CreateRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseCreateRoomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/baseCreateRoomRequest"
            }
          }
        ],
        "tags": [
          "RoomService"
        ]
      }
    },
...
    "/v1/rooms/{roomId}": {
      "get": {
        "operationId": "RoomService_GetRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseGetRoomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "RoomService"
        ]
      },
      "delete": {
        "operationId": "RoomService_DeleteRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseMessageStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "RoomService"
        ]
      },
      "put": {
        "operationId": "RoomService_UpdateRoom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/baseUpdateRoomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RoomServiceUpdateRoomBody"
            }
          }
        ],
        "tags": [
          "RoomService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "baseCreateBookingRequest": {
      "type": "object",
      "properties": {
        "tenantId": {
          "type": "integer",
          "format": "int32"
        },
        "roomId": {
          "type": "integer",
          "format": "int32"
        },
        "checkInDate": {
          "type": "string",
          "format": "date-time",
          "title": "Renamed from start_date"
        },
        "checkOutDate": {
          "type": "string",
          "format": "date-time",
          "title": "Renamed from end_date"
        }
      },
      "title": "Booking Service Messages"
    },
    "baseCreateBookingResponse": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/baseBooking"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseCreateInvoiceRequest": {
      "type": "object",
      "properties": {
        "bookingId": {
          "type": "integer",
          "format": "int32"
        },
        "paymentMethod": {
          "type": "string"
        },
        "dueDate": {
          "type": "string",
          "format": "date-time"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      },
      "title": "Payment Service Messages"
    },
    "baseCreateInvoiceResponse": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/baseInvoice"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseCreateRoomRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "classificationId": {
          "type": "integer",
          "format": "int32"
        },
        "rentalTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "capacity": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "amenities": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/baseRoomAmenityInput"
          }
        }
      }
    },
    "baseCreateRoomResponse": {
      "type": "object",
      "properties": {
        "room": {
          "$ref": "#/definitions/baseRoom"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseGetBookingResponse": {
      "type": "object",
      "properties": {
        "booking": {
          "$ref": "#/definitions/baseBooking"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseGetBookingsResponse": {
      "type": "object",
      "properties": {
        "bookings": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/baseBooking"
          }
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseGetInvoiceResponse": {
      "type": "object",
      "properties": {
        "invoice": {
          "$ref": "#/definitions/baseInvoice"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseGetInvoicesResponse": {
      "type": "object",
      "properties": {
        "invoices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/baseInvoice"
          }
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseGetIssueResponse": {
      "type": "object",
      "properties": {
        "issue": {
          "$ref": "#/definitions/baseIssue"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseGetIssuesResponse": {
      "type": "object",
      "properties": {
        "issues": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/baseIssue"
          }
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseGetRoomResponse": {
      "type": "object",
      "properties": {
        "room": {
          "$ref": "#/definitions/baseRoom"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
    "baseGetRoomsResponse": {
      "type": "object",
      "properties": {
        "rooms": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/baseRoom"
          }
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseMessageStatusResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "title": "===================================================================\nCommon Messages\n==================================================================="
    },
...
    "basePaymentCallbackRequest": {
      "type": "object",
      "properties": {
        "transactionStatus": {
          "type": "string"
        },
        "orderId": {
          "type": "string"
        },
        "paymentType": {
          "type": "string"
        },
        "fraudStatus": {
          "type": "string"
        },
        "statusCode": {
          "type": "string"
        },
        "grossAmount": {
          "type": "string",
          "title": "You can add more fields from the Midtrans callback payload as needed\n For simplicity, you might also choose to receive the raw JSON string\n string raw_payload = 7;"
        }
      }
    },
...
    "baseReportIssueRequest": {
      "type": "object",
      "properties": {
        "bookingId": {
          "type": "integer",
          "format": "int32"
        },
        "reportedByUserId": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "fileName": {
          "type": "string",
          "title": "Added for image upload"
        },
        "fileType": {
          "type": "string",
          "title": "Added for image upload"
        },
        "content": {
          "type": "string",
          "format": "byte",
          "title": "Added for image upload"
        }
      },
      "title": "Issue Service Messages"
    },
    "baseReportIssueResponse": {
      "type": "object",
      "properties": {
        "issue": {
          "$ref": "#/definitions/baseIssue"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "baseUpdateRoomResponse": {
      "type": "object",
      "properties": {
        "room": {
          "$ref": "#/definitions/baseRoom"
        },
        "status": {
          "$ref": "#/definitions/baseMessageStatusResponse"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}